+++ marketplace/applications/models.py
@@ -32,6 +32,7 @@ class App(AppTypeBaseModel):
     project_uuid = models.UUIDField("Project UUID")
     platform = models.CharField(choices=PLATFORM_CHOICES, max_length=2)
     flow_object_uuid = models.UUIDField(null=True, unique=True)
+    configured = models.BooleanField(default=False)
 
     class Meta:
         verbose_name = _("App")
@@ -85,6 +86,8 @@ class AppTypeAsset(AppTypeBaseModel):
 
 
 class AppTypeFeatured(AppTypeBaseModel):
+    priority = models.PositiveSmallIntegerField(default=0)
+
     class Meta:
         verbose_name = "AppType Featured"
         verbose_name_plural = "AppType Featureds"
@@ -97,5 +100,5 @@ class AppTypeFeatured(AppTypeBaseModel):
 
     @classmethod
     def get_apptype_featureds(cls) -> Generator[None, None, "AppType"]:
-        for featured in cls.objects.all():
+        for featured in cls.objects.order_by("priority"):
             yield featured.apptype